added functionality for changing style on the first post in a thread

+++ src/forum/kinds/Forum.PostContainer.js
@@ -5,11 +5,12 @@ enyo.kind({
     tag: "div",
     
     components: [
-		{name: "threadid", tag: "p"},
-		{name: "posts", tag: "ul"}
-	    ],
-	    published: {
-		threadid: "Default id"
+	{name: "threadid", tag: "p"},
+	{name: "threadPost", tag: "ul"},
+	{name: "posts", tag: "ul"}
+    ],
+    published: {
+	threadid: "Default id"
     },
     
     newPost: function (){
@@ -17,8 +18,16 @@ enyo.kind({
     },
 
     create: function () {
-		this.inherited(arguments);
-		this.populate();
+	this.inherited(arguments);
+	this.createComponent({
+	    kind: "ThreadPost",
+	    container: this.$.threadPost,
+	    text: this.thread.content,
+	    userid: this.thread.user,
+	    dbparent: this.thread,
+	    post: this.thread.content
+	});
+	this.populate();
     },
 
     populate: function () {
@@ -30,19 +39,23 @@ enyo.kind({
     },
     
     gotPosts: function(list) {
-		this.$.posts.destroyClientControls();
-		enyo.forEach(list.items(), function(post) {
-		   // console.log(post);
-		    this.createComponent({
-			kind: "ForumPost",
-			container: this.$.posts,
-			text: post.text,
-			userid: post.user,
-			dbparent: this.thread,
-			post: post
-		    });
-		}, this);
-		this.$.posts.render();
+	if ( this.destroyed ) {
+	    list.close();
+	    return;
+	}
+	this.$.posts.destroyClientControls();
+	enyo.forEach(list.items(), function(post) {
+	    // console.log(post);
+	    this.createComponent({
+		kind: "ForumPost",
+		container: this.$.posts,
+		text: post.text,
+		userid: post.user,
+		dbparent: this.thread,
+		post: post
+	    });
+	}, this);
+	this.$.posts.render();
     }
 
 });
